% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_termlist}
\alias{check_termlist}
\title{Check Termlist}
\usage{
check_termlist(.tab, .fun_std = NULL)
}
\arguments{
\item{.tab}{A Dataframe with at least 1 column (term: Term)}

\item{.fun_std}{A
function to standardize Strings. Default = NULL (no standardization use)}
}
\value{
Dataframe and Messages/Warnings
}
\description{
Check Termlist
}
\examples{
check_termlist(table_termlist_short, NULL)
check_termlist(table_termlist_short, string_standardization)

termlist <- tibble::add_row(table_termlist_short, term = "Language-Processing")
check_termlist(termlist, NULL)
check_termlist(termlist, string_standardization)
}
