% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_context}
\alias{get_context}
\title{Get Context}
\usage{
get_context(
  .position,
  .document,
  .n = NA,
  .context = c("word", "sentence", "paragraph", "page"),
  .vars = c("none", "sentence", "paragraph", "page")
)
}
\arguments{
\item{.position}{A Dataframe produced by position_count()}

\item{.document}{A Dataframe produced by prep_document()}

\item{.n}{Number of context token to retrieve, default = NA, which means no token context is retrieved}

\item{.context}{either "word", "sentence", "paragraph", or "page"}

\item{.vars}{Any combination of "sentence", "paragraph", and "page" (get the specific locations of hits)\cr
If not specified, no variable context is retrieved}
}
\value{
A dataframe
}
\description{
Get Context
}
\examples{
termlist_short <- prep_termlist(table_termlist_short, string_standardization, TRUE, tid)
document       <- prep_document(table_document_short, string_standardization)

tab_pos_short <- position_count(
  termlist_short, document, sen_id, .cache_terms = TRUE, .tab_pos = NULL
)

tab_context_word     <- get_context(tab_pos_short, document, 5, "word")
tab_context_sentence <- get_context(tab_pos_short, document, 1, "sentence")
tab_context_vars     <- get_context(tab_pos_short, document, .vars = "sentence")
}
