% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{h_dependencies_termlist}
\alias{h_dependencies_termlist}
\title{Helper Function: Get Term Dependencies}
\usage{
h_dependencies_termlist(.termlist)
}
\arguments{
\item{.termlist}{A Datframe produced by h_prep_termlist()}
}
\value{
A Dataframe
}
\description{
Helper Function: Get Term Dependencies
}
