% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{h_position_count}
\alias{h_position_count}
\title{Helper Function for position_count()}
\usage{
h_position_count(.row_terms, .document)
}
\arguments{
\item{.row_terms}{A one row Dataframe prepared by prep_termlist()}

\item{.document}{A tokenized Dataframe prepared by prep_document()}
}
\value{
A Dataframe
}
\description{
Helper Function for position_count()
}
