% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{h_prep_termlist}
\alias{h_prep_termlist}
\title{Helper Function: Prepare termlist}
\usage{
h_prep_termlist(.tab, .fun_std = NULL)
}
\arguments{
\item{.tab}{A Dataframe with a column 'term'}

\item{.fun_std}{standardization function}
}
\value{
A Datframe
}
\description{
Helper Function: Prepare termlist
}
