% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{position_count_old}
\alias{position_count_old}
\title{Retrieve Position of Terms in Documents}
\usage{
position_count_old(.termlist, .document, ...)
}
\arguments{
\item{.termlist}{A term list prepared by prep_termlist()}

\item{.document}{A tokenized Dataframe prepared by prep_document()}

\item{...}{Any number of Columns that mark a separator of tokens (e.g. a sentence)}
}
\value{
A Dataframe
}
\description{
Retrieve Position of Terms in Documents
}
