% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prep_termlist}
\alias{prep_termlist}
\title{Prepare Termlist}
\usage{
prep_termlist(.tab, .fun_std = NULL, .get_dep = FALSE)
}
\arguments{
\item{.tab}{A dataframe with at least 1 column (term: Term)}

\item{.fun_std}{A
Function to standardize strings. Default = NULL (no standardization use)}

\item{.get_dep}{Get term list dependencies (default = FALSE)}
}
\value{
A dataframe with one column including the terms for which the function should look for. Terms can be any combination of alphanumeric values, including n-grams. For example, search for "1970", "reporting" and "reporting frequency" is possible alike.
}
\description{
Prepare Termlist
}
\examples{

prep_termlist(table_termlist_short, NULL)
prep_termlist(table_termlist_short, string_standardization)
prep_termlist(table_termlist_short, string_standardization, TRUE)
}
