% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prep_termlist_old}
\alias{prep_termlist_old}
\title{Prepare Term List}
\usage{
prep_termlist_old(.tab, .fun_std = NULL)
}
\arguments{
\item{.tab}{A Dataframe with at least 2 columns (tid: Term ID, term: Term)}

\item{.fun_std}{A function to standardize Strings. Default = NULL (no standardization use)}
}
\value{
A Dataframe with a term list column
}
\description{
Prepare Term List
}
\examples{

termlist <- table_termlist_short
prep_termlist(termlist)
}
