% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{string_standardization}
\alias{string_standardization}
\title{Basic String Standardization}
\usage{
string_standardization(.str)
}
\arguments{
\item{.str}{A string}
}
\value{
A string
}
\description{
Basic String Standardization
}
\examples{
string_standardization("  TesT String")
}
