% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentations.R
\docType{data}
\name{table_termlist_long}
\alias{table_termlist_long}
\title{Example Term List (Long)}
\format{
A data frame 12 rows and 1 column
\describe{
\item{term}{A Term}
}
}
\usage{
table_termlist_long
}
\description{
A dataset containing example terms to be searched within a document
}
\keyword{datasets}
