% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentations.R
\docType{data}
\name{table_termlist_short}
\alias{table_termlist_short}
\title{Example Term List (Short)}
\format{
A data frame 7 rows and 1 column
\describe{
\item{term}{A Term}
}
}
\usage{
table_termlist_short
}
\description{
A dataset containing example terms to be searched within a document
}
\keyword{datasets}
