% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentations.R
\docType{data}
\name{test_document}
\alias{test_document}
\title{Wikipedia Article about Natural Language Processing (NLP)}
\format{
A data frame one row and two columns
\describe{
\item{doc_id}{A unique identifier of the document}
\item{text}{Text of the document}
}
}
\usage{
test_document
}
\description{
A dataset containing an example text from wikipedia
\href{https://en.wikipedia.org/wiki/Natural_language_processing}{Wikipedia}
}
\keyword{datasets}
