% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentations.R
\docType{data}
\name{test_termlist}
\alias{test_termlist}
\title{Example Term List}
\format{
A data frame one row and two columns
\describe{
\item{tid}{A unique identifier of the term (tid: Term ID)}
\item{term}{A Term}
}
}
\usage{
test_termlist
}
\description{
A dataset containing example terms to be searched within a document
}
\keyword{datasets}
