% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_termlist}
\alias{check_termlist}
\title{Check Termlist}
\usage{
check_termlist(.tab, .fun_std = NULL)
}
\arguments{
\item{.tab}{A Dataframe with at least 1 column (term: Term)}

\item{.fun_std}{A Function to standardize Strings.\cr
Default = NULL (no standardization use)\cr
Build-in Function .fun_std = string_standardization, can be used.}
}
\value{
Case 1: No Duplicates: A Message with Dependency Factor (unique terms / all terms)\cr
Case 2: A Warning Message and a Dataframe with duplicated values (AFTER STANDARDIZATION)
}
\description{
Check Termlist
}
\examples{
# Example w/o standardization
check_termlist(table_termlist_short, NULL)
# Example with standardization
check_termlist(table_termlist_short, string_standardization)

# Example w/o standardization and duplicated term
termlist <- tibble::add_row(table_termlist_short, term = "Language-Processing")
check_termlist(termlist, NULL)

# Example with standardization and duplicated term
check_termlist(termlist, string_standardization)
}
