% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_seq_in_seq}
\alias{find_seq_in_seq}
\title{Find a sequence in another sequence}
\usage{
find_seq_in_seq(.seq_find, .seq_base)
}
\arguments{
\item{.seq_find}{The sequence you want to find in another sequence}

\item{.seq_base}{The sequence to be searched}
}
\value{
An integer vector (vector has length zero if sequence is not found)
}
\description{
Find a sequence in another sequence
}
\examples{
# Find an integer sequence
find_seq_in_seq(2:10, c(1:20, 1:20))

# Find a character sequence
find_seq_in_seq(c("C", "D"), LETTERS)

# No sequence found
find_seq_in_seq(c("R", "D"), LETTERS)

}
