% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prep_termlist}
\alias{prep_termlist}
\title{Prepare Termlist}
\usage{
prep_termlist(.tab, .fun_std = NULL)
}
\arguments{
\item{.tab}{A Dataframe with at least 1 column (term: Term)}

\item{.fun_std}{A
function to standardize Strings. Default = NULL (no standardization use)}
}
\value{
A Dataframe
}
\description{
Prepare Termlist
}
\examples{

prep_termlist(table_termlist_short, NULL)
prep_termlist(table_termlist_short, string_standardization)

}
