% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prep_termlist}
\alias{prep_termlist}
\title{Prepare Term List}
\usage{
prep_termlist(.tab)
}
\arguments{
\item{.tab}{A Dataframe with at least 2 columns (tid: Term ID, term: Term)}
}
\value{
A Dataframe with a term list column
}
\description{
Prepare Term List
}
\examples{

termlist <- test_termlist
prep_termlist(termlist)
}
